requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.4
beir==2.1.0
bitsandbytes==0.46.0
blis==1.3.0
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docutils==0.21.2
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.16.1
flash-attn==2.7.0.post2
fonttools==4.55.6
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.10.0
gradio==5.6.0
gradio_client==1.4.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.3.1
ipykernel==6.29.5
ipython==8.32.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==3.12.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.5.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.77.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.60.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
preshed==3.0.9
prompt-toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.1
PyMuPDF==1.26.5
pynvml==12.0.0
pyparsing==3.2.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytrec_eval-terrier==0.5.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rouge==1.0.1
rsa==4.9.1
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
SecretStorage==3.3.3
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.1.0
twine==6.0.1
typeguard==4.3.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
websockets==11.0.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zopfli==0.2.3.post1
